leftwm(bin) fully moved into leftwm-watchdog crate
