+++ leftwm-watchdog/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "leftwm"
+name = "leftwm-watchdog"
 version = "0.5.1"
 rust-version = "1.70.0"
 authors = ["Lex Childs <lexchilds@gmail.com>"]
@@ -13,6 +13,11 @@ description = "A window manager for Adventurers"
 
 [dependencies]
 clap = { version = "4.0.0", features = ["cargo"] }
+dirs-next = "2.0.0"
+nix = { version = "0.27.1", features = ["fs", "signal"] }
+signal-hook = "0.3.4"
+thiserror = "1.0.30"
+xdg = "2.2.0"
 
 [dev-dependencies]
 tempfile = "3.2.0"
@@ -24,4 +29,5 @@ lefthk = []
 slow-dm-fix = []
 
 [[bin]]
-
+name = "leftwm"
+path = "src/main.rs"
