+++ leftwm-watchdog/src/main.rs
@@ -12,8 +12,7 @@ use std::sync::{
     Arc,
 };
 
-mod util;
-use util::{self, Nanny};
+mod utils;
 
 type Subcommand<'a> = &'a str;
 type SubcommandArgs = Vec<String>;
@@ -81,13 +80,14 @@ fn print_help_page() {
     };
 
     command!()
+        .bin_name("leftwm")
         .about(
             "Starts LeftWM if no arguments are supplied. If a subcommand is given, executes the \
              the corresponding leftwm program, e.g. 'leftwm theme' will execute 'leftwm-theme', if \
              it is installed.",
         )
         .subcommands(subcommands)
-        .help_template(leftwm::utils::get_help_template())
+        .help_template(utils::get_help_template())
         .print_help()
         .unwrap();
 }
@@ -160,7 +160,7 @@ fn start_leftwm() {
     set_env_vars();
 
     // Boot everything WM agnostic or LeftWM related in ~/.config/autostart
-    let mut children = Nanny::autostart();
+    let mut children = utils::autostart();
 
     let flag = get_sigchld_flag();
 
@@ -173,7 +173,7 @@ fn start_leftwm() {
 
         while session_is_running(&mut leftwm_session) {
             // remove all child processes which finished
-            children.remove_finished_children();
+            utils::remove_finished_children(&mut children);
 
             while is_suspending(&flag) {
                 nix::unistd::pause();
@@ -253,7 +253,7 @@ fn kill_lefthk_session(lefthk_session: &mut Child) {
 /// example-description.
 fn get_sigchld_flag() -> Arc<AtomicBool> {
     let flag = Arc::new(AtomicBool::new(false));
-    child_process::register_child_hook(flag.clone());
+    utils::register_child_hook(flag.clone());
 
     flag
 }
