+++ leftwm/Cargo.toml
@@ -16,7 +16,6 @@ anyhow = "1.0.48"
 clap = { version = "4.0.0", features = ["cargo"] }
 const_format = "0.2.26"
 once_cell = "1.13.0"
-dirs-next = "2.0.0"
 futures = "0.3.21"
 git-version = "0.3.5"
 lefthk-core = { version = '0.2', optional = true }
@@ -25,7 +24,6 @@ leftwm-macros = {path = "../leftwm-macros", version = '0.5.0'}
 leftwm-layouts = "0.8.4"
 liquid = "0.26.0"
 mio = "0.8.0"
-nix = {version = "0.27.1", features = ["fs", "signal"]}
 regex = "1"
 ron = "0.8.0"
 serde = { version = "1.0.104", features = ["derive", "rc"] }
